+++ site/plugin-tdoc/md-to-vue.js
@@ -3,7 +3,7 @@ import fs from 'fs';
 import path from 'path';
 import matter from 'gray-matter';
 import camelCase from 'camelcase';
-import { compileUsage } from '../../src/_common/docs/compile';
+import { compileUsage, getGitTimestamp } from '../../src/_common/docs/compile';
 
 import testCoverage from '../test-coverage';
 
@@ -13,8 +13,8 @@ const DEFAULT_TABS = [
   { tab: 'design', name: '指南' },
 ];
 
-export default function mdToVue(options) {
-  const mdSegment = customRender(options);
+export default async function mdToVue(options) {
+  const mdSegment = await customRender(options);
   const { demoDefsStr, demoCodesDefsStr, demoInstallStr, demoCodeInstallStr } = options;
 
   let coverage = '';
@@ -57,7 +57,7 @@ export default function mdToVue(options) {
             : `<div name="DOC" class="${mdSegment.docClass}">${mdSegment.docMd}</div>`
         }
         <div style="margin-top: 48px;">
-          <td-doc-history time="${mdSegment.lastUpdated}"></td-doc-history>
+          <td-doc-history :time="lastUpdated" :key="lastUpdated"></td-doc-history>
         </div>
         <td-doc-footer slot="doc-footer"></td-doc-footer>
       </td-doc-content>
@@ -83,6 +83,10 @@ export default function mdToVue(options) {
         },
 
         computed: {
+          lastUpdated() {
+            if (this.tab === 'design') return ${mdSegment.designDocLastUpdated};
+            return ${mdSegment.lastUpdated};
+          },
           tab: {
             get() {
               return this.$route.query.tab || 'demo';
@@ -124,8 +128,9 @@ export default function mdToVue(options) {
 }
 
 // 解析 markdown 内容
-function customRender({ source, file, md }) {
+async function customRender({ source, file, md }) {
   const { content, data } = matter(source);
+  const lastUpdated = (await getGitTimestamp(file)) || Math.round(fs.statSync(file).mtimeMs);
   // console.log('data', data);
 
   // md top data
@@ -140,7 +145,8 @@ function customRender({ source, file, md }) {
     apiFlag: /#+\s*API/,
     docClass: '',
     usage: null,
-    lastUpdated: Math.round(fs.statSync(file).mtimeMs),
+    lastUpdated,
+    designDocLastUpdated: lastUpdated,
     ...data,
   };
 
@@ -195,6 +201,10 @@ function customRender({ source, file, md }) {
     const designDocPath = path.resolve(__dirname, `../../src/_common/docs/web/design/${componentName}.md`);
 
     if (fs.existsSync(designDocPath)) {
+      const designDocLastUpdated =
+        (await getGitTimestamp(designDocPath)) || Math.round(fs.statSync(designDocPath).mtimeMs);
+      mdSegment.designDocLastUpdated = designDocLastUpdated;
+
       const designMd = fs.readFileSync(designDocPath, 'utf-8');
       mdSegment.designMd = md.render.call(md, `${pageData.toc ? '[toc]\n' : ''}${designMd}`).html;
     } else {