+++ src/avatar/avatar.tsx
@@ -4,6 +4,7 @@ import { renderContent, renderTNodeJSX } from '../utils/render-tnode';
 import { AvatarGroupInstance } from './instance';
 import { Styles } from '../common';
 import { getClassPrefixMixins } from '../config-provider/config-receiver';
+import Image from '../image';
 import mixins from '../utils/mixins';
 
 const classPrefixMixins = getClassPrefixMixins('avatar');
@@ -22,7 +23,6 @@ export default mixins(Vue as VueConstructor<AvatarInstance>, classPrefixMixins).
     return {
       isImgExist: true,
       gap: 4,
-      sizeValue: '',
       scale: '',
     };
   },
@@ -32,6 +32,10 @@ export default mixins(Vue as VueConstructor<AvatarInstance>, classPrefixMixins).
     },
   },
   computed: {
+    // 即使 Vue2 的 inject 非响应式，但放在 computed 中依然正常响应
+    sizeValue(): string {
+      return this.size || this.avatarGroup?.size;
+    },
     customAvatarSize(): Styles {
       return this.isCustomSize()
         ? {
@@ -57,19 +61,17 @@ export default mixins(Vue as VueConstructor<AvatarInstance>, classPrefixMixins).
   },
 
   mounted() {
-    const { avatarGroup } = this;
-    this.sizeValue = this.size || avatarGroup?.size;
     this.$nextTick(() => {
       this.setScaleParams();
     });
   },
 
   methods: {
-    handleImgLoadError() {
-      const { onError, hideOnLoadFailed } = this.$props;
+    handleImgLoadError(context: { e: Event }) {
+      const { onError, hideOnLoadFailed } = this;
       this.isImgExist = !hideOnLoadFailed;
-      onError && onError();
-      this.$emit('error');
+      this.$emit('error', context);
+      onError && onError(context);
     },
     setScaleParams() {
       const avatar = this.$refs.avatar as HTMLElement;
@@ -115,10 +117,23 @@ export default mixins(Vue as VueConstructor<AvatarInstance>, classPrefixMixins).
     }
 
     if (image && this.isImgExist) {
-      content = <img style={{ ...this.customImageSize }} src={image} alt={alt} onError={this.handleImgLoadError}></img>;
+      content = (
+        <Image
+          style={{ ...this.customImageSize }}
+          src={image}
+          alt={alt}
+          props={this.imageProps}
+          error={!this.$scopedSlots.error && !this.imageProps?.error ? ' ' : undefined}
+          loading={!this.$scopedSlots.loading && !this.imageProps?.loading ? ' ' : undefined}
+          scopedSlots={this.$scopedSlots}
+          on={{
+            error: this.handleImgLoadError,
+          }}
+        ></Image>
+      );
     }
     return (
-      <div ref="avatar" class={avatarClass} style={{ ...this.customAvatarSize }}>
+      <div ref="avatar" class={avatarClass} style={this.customAvatarSize}>
         {content}
       </div>
     );