+++ src/common.ts
@@ -31,6 +31,8 @@ export interface UploadDisplayDragEvents {
   dragover?: (event: DragEvent) => void;
   dragleave?: (event: DragEvent) => void;
 }
+
+export type ImageEvent = Event;
 /** 通用全局类型 */
 
 export type OptionData = {