+++ src/input/input.tsx
@@ -100,7 +100,7 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
           [this.commonSizeClassName[this.size]]: this.size !== 'medium',
           [this.commonStatusClassName.disabled]: this.tDisabled,
           [this.commonStatusClassName.focused]: this.focused,
-          [`${this.classPrefix}-is-${this.tStatus}`]: this.tStatus,
+          [`${this.classPrefix}-is-${this.tStatus}`]: this.tStatus && this.tStatus !== 'default',
           [`${this.classPrefix}-align-${this.align}`]: this.align !== 'left',
           [`${this.classPrefix}-is-disabled`]: this.tDisabled,
           [`${this.classPrefix}-is-readonly`]: this.readonly,
@@ -168,6 +168,12 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
       handler(val) {
         this.inputValue = this.format ? this.format(val) : val;
         this.preValue = this.inputValue;
+
+        // limit props value
+        const newVal = this.getValueByLimitNumber(val);
+        if (newVal !== val && this.type !== 'number') {
+          emitEvent<Parameters<TdInputProps['onChange']>>(this, 'change', newVal, { trigger: 'initial' });
+        }
       },
       immediate: true,
     },
@@ -238,11 +244,11 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
 
       // 插槽名称为中划线
       if (this.$scopedSlots[kebabCase(iconType)]) {
-        return this.$scopedSlots[kebabCase(iconType)](null);
+        return this.$scopedSlots[kebabCase(iconType)](h);
       }
       // 插槽名称为驼峰
       if (this.$scopedSlots[camelCase(iconType)]) {
-        return this.$scopedSlots[camelCase(iconType)](null);
+        return this.$scopedSlots[camelCase(iconType)](h);
       }
 
       return null;
@@ -289,22 +295,24 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
       }: any = e;
       emitEvent<Parameters<TdInputProps['onKeypress']>>(this, 'keypress', value, { e });
     },
-    onHandlePaste(e: ClipboardEvent) {
+    handlePaste(e: ClipboardEvent) {
       if (this.tDisabled) return;
       // @ts-ignore
       const clipData = e.clipboardData || window.clipboardData;
       emitEvent<Parameters<TdInputProps['onPaste']>>(this, 'paste', { e, pasteValue: clipData?.getData('text/plain') });
     },
+
     onHandleMousewheel(e: WheelEvent) {
       emitEvent<Parameters<TdInputProps['onWheel']>>(this, 'wheel', { e });
     },
+
     emitPassword() {
       const { renderType } = this;
       const toggleType = renderType === 'password' ? 'text' : 'password';
       this.renderType = toggleType;
     },
     emitClear(e: MouseEvent) {
-      emitEvent<Parameters<TdInputProps['onChange']>>(this, 'change', '', { e });
+      emitEvent<Parameters<TdInputProps['onChange']>>(this, 'change', '', { e, trigger: 'clear' });
       emitEvent<Parameters<TdInputProps['onClear']>>(this, 'clear', { e });
     },
     emitFocus(e: FocusEvent) {
@@ -327,7 +335,8 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
         currentTarget: { value },
       }: any = e;
       this.composingRefValue = value;
-      this?.onCompositionstart?.(value, { e });
+      this.$emit('compositionstart', value, { e });
+      this.onCompositionstart?.(value, { e });
     },
     compositionendHandler(e: CompositionEvent) {
       const {
@@ -338,11 +347,13 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
         this.handleInput(e);
       }
       this.composingRefValue = '';
-      this?.onCompositionend?.(value, { e });
+      this.$emit('compositionend', value, { e });
+      this.onCompositionend?.(value, { e });
     },
     onRootClick(e: MouseEvent) {
       (this.$refs.inputRef as HTMLInputElement)?.focus();
-      this.$emit('click', e);
+      this.$emit('click', { e });
+      this.onClick?.({ e });
     },
     handleInput(e: InputEvent | CompositionEvent) {
       this.preValue = this.inputValue + e.data;
@@ -356,7 +367,7 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
         if (this.type !== 'number') {
           val = this.getValueByLimitNumber(val);
         }
-        emitEvent<Parameters<TdInputProps['onChange']>>(this, 'change', val, { e } as { e: MouseEvent | InputEvent });
+        emitEvent<Parameters<TdInputProps['onChange']>>(this, 'change', val, { e, trigger: 'input' });
         // 受控，重要，勿删 input无法直接实现受控
         if (!this.isIE) {
           const inputRef = this.$refs.inputRef as HTMLInputElement;
@@ -374,12 +385,12 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
 
     onInputMouseenter(e: MouseEvent) {
       this.mouseEvent(true);
-      this.onMouseenter?.({ e });
+      emitEvent<Parameters<TdInputProps['onMouseenter']>>(this, 'mouseenter', { e });
     },
 
     onInputMouseleave(e: MouseEvent) {
       this.mouseEvent(false);
-      this.onMouseleave?.({ e });
+      emitEvent<Parameters<TdInputProps['onMouseleave']>>(this, 'mouseleave', { e });
     },
 
     updateInputWidth() {
@@ -414,17 +425,19 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
   },
 
   render(h: CreateElement): VNode {
-    const inputEvents = getValidAttrs({
+    const inputEvents = {
       focus: this.emitFocus,
       blur: this.formatAndEmitBlur,
       keydown: this.handleKeydown,
       keyup: this.handleKeyUp,
       keypress: this.handleKeypress,
-      paste: this.onHandlePaste,
+      paste: this.handlePaste,
+      compositionstart: this.compositionstartHandler,
+      compositionend: this.compositionendHandler,
       // input的change事件是失去焦点或者keydown的时候执行。这与api定义的change不符，所以不做任何变化。
       // eslint-disable-next-line @typescript-eslint/no-empty-function
       change: () => {},
-    });
+    };
 
     const wrapperAttrs = omit(this.$attrs, Object.keys(this.inputAttrs));
     const wrapperEvents = omit(this.$listeners, [...Object.keys(inputEvents), 'input', 'paste']);
@@ -482,11 +495,12 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
         [`${this.componentName}--suffix`]: suffixIcon || suffixContent,
       },
     ];
+
     const inputNode = (
       <div
         class={classes}
-        onClick={this.onRootClick}
         {...{ attrs: wrapperAttrs, on: wrapperEvents }}
+        onClick={this.onRootClick}
         onMouseenter={this.onInputMouseenter}
         onMouseleave={this.onInputMouseleave}
         onwheel={this.onHandleMousewheel}
@@ -497,13 +511,12 @@ export default mixins(getConfigReceiverMixins<InputInstance, InputConfig>('input
         {labelContent}
         {this.showInput && (
           <input
-            {...{ attrs: this.inputAttrs, on: inputEvents }}
+            attrs={this.inputAttrs}
+            on={inputEvents}
             ref="inputRef"
             class={`${this.componentName}__inner`}
             value={this.composingRef ? this.composingRefValue : this.inputValue}
             onInput={this.handleInput}
-            onCompositionstart={this.compositionstartHandler}
-            onCompositionend={this.compositionendHandler}
           />
         )}
         {this.autoWidth && (