+++ src/popup/container.tsx
@@ -1,7 +1,8 @@
 import Vue from 'vue';
 import raf from 'raf';
+import { PropType } from '@vue/composition-api';
 import { getAttach, removeDom } from '../utils/dom';
-import props from './props';
+import { TdPopupProps } from './type';
 
 function isContentRectChanged(rect1: DOMRectReadOnly, rect2: DOMRectReadOnly) {
   if (!rect1 || !rect2) return;
@@ -60,7 +61,13 @@ export default Vue.extend({
   props: {
     parent: Object,
     visible: Boolean,
-    attach: props.attach,
+    // support attach to current node when current is equal to `CURRENT_NODE`
+    attach: [Function] as PropType<
+      () => {
+        attach: TdPopupProps['attach'];
+        current: HTMLElement;
+      }
+    >,
   },
   data() {
     return {
@@ -91,7 +98,7 @@ export default Vue.extend({
       this.content = new (this.$root.constructor as any)({
         parent,
         render() {
-          return <div>{parent.$slots.content}</div>;
+          return parent.$slots.content;
         },
         mounted() {
           parent.$emit('contentMounted');
@@ -110,8 +117,10 @@ export default Vue.extend({
           removeDom(elm);
         },
       });
+      const { attach, current } = this.attach();
+      const currentAttach = attach === 'CURRENT_NODE' ? current : attach;
       // @ts-ignore
-      getAttach(this.attach, this.$refs?.triggerRef?.$el).appendChild(elm);
+      getAttach(currentAttach, this.$refs?.triggerRef?.$el).appendChild(elm);
       this.content.$mount(elm.children[0]);
     },
     unmountContent() {